requirements.txt
pydub==0.25.1
moviepy==2.2.1
numpy==2.3.3
decorator==5.2.1
ffmpeg-python==0.2.0
audioop-lts==0.2.2